LayoutTests/platform/chromium-cg-mac/fast/dom/timer-clear-interval-in-handler-and-generate-error-expected.txt

CONSOLE MESSAGE: line 10: Uncaught Error: NOT_FOUND_ERR: DOM Exception 8
